+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [0.1.1-preview.1] - 2021-11-23
+- Change minimum requirement Unity version to 2020.3
+- Change dependency Text Mesh Pro to version 3.0.6
+- Added DSC.UI.Editor assembly definition.
+- Added ContextUI class.
+
 ## [0.1.0-preview.1] - 2021-01-17
 - Change minimum requirement Unity version to 2020.2
 - Change dependency Text Mesh Pro to version 3.0.3